Color Detection System
Hi,  I have created a Color Detection System, which is the process of detecting the name of any color. Our brain then recognizes the color.  I used the somewhat same strategy to detect color names. In this color detection Python project, I was building an application through which you can automatically get the name of the color by clicking on them. So for this,  I have a data file that contains the color name and its values. Then I calculated the distance from each color and find the shortest one.  Make sure to give an image path using ‘-i’ argument. If the image is in another directory, then you need to give full path of the image:

python color_detection.py -i <add your image path here>

I think it's a new idea for machine learning projects. if you like it then accept my request and merge with your data.